Split a long go-file into short files for easy maintanence
